refactor(charts): extract pattern-fill methods into ChartPatternMixin
Move _pattern_color_count, _pattern_defs, _pattern_id, _category_fill,
and _filled_outline_attrs out of the Chart base class into a new
_chart_patterns module. Bodies are unchanged; the mixin is added as the
first base so the methods resolve on Chart identically.
